# IOBuffer back to Ruby