--- a/ext/puma_http11/PumaHttp11Service.java
+++ b/ext/puma_http11/PumaHttp11Service.java
@@ -1,7 +1,7 @@
 package puma;
 
 import java.io.IOException;
-        
+
 import org.jruby.Ruby;
 import org.jruby.runtime.load.BasicLibraryService;
 
@@ -9,10 +9,9 @@ import org.jruby.puma.Http11;
 import org.jruby.puma.IOBuffer;
 import org.jruby.puma.MiniSSL;
 
-public class PumaHttp11Service implements BasicLibraryService { 
+public class PumaHttp11Service implements BasicLibraryService {
     public boolean basicLoad(final Ruby runtime) throws IOException {
         Http11.createHttp11(runtime);
-        IOBuffer.createIOBuffer(runtime);
         MiniSSL.createMiniSSL(runtime);
         return true;
     }