… ext/puma_http11/io_buffer.c
… ext/puma_http11/org/jruby/puma/IOBuffer.java