File under "weird idea".