@eregon 2019-09-23
This looks great to me:
-155 lines of C
-72 lines of Java
+17 lines of Ruby

I guess the main concern is: how does it affect performance?