@kares 2019-10-07
a bit ugly that `IOBuffer` now has a "full" `String` API but I guess its not a user API anyway ...