@nateberkopec 2019-10-13
This branch, extra large request body:
```
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     1.96ms    1.02ms  16.99ms   72.38%
    Req/Sec     1.04k    98.31     1.22k    79.08%
  Latency Distribution
     50%    1.90ms
     75%    2.52ms
     90%    3.24ms
     99%    4.86ms
  124562 requests in 1.00m, 127.62GB read
Requests/sec:   2074.35
Transfer/sec:      2.13GB
```

Master, large request body:
```
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     1.74ms    0.86ms  16.93ms   71.69%
    Req/Sec     1.17k    66.38     1.29k    85.25%
  Latency Distribution
     50%    1.68ms
     75%    2.26ms
     90%    2.81ms
     99%    3.99ms
  139603 requests in 1.00m, 143.03GB read
Requests/sec:   2325.64
Transfer/sec:      2.38GB
```

This branch, hello world:
```
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   242.78us  137.14us   2.63ms   97.29%
    Req/Sec     8.67k   661.24     9.04k    96.52%
  Latency Distribution
     50%  224.00us
     75%  258.00us
     90%  290.00us
     99%    1.06ms
  173345 requests in 10.11s, 12.56MB read
Requests/sec:  17154.11
Transfer/sec:      1.24MB
```

Master, hello world:

```
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   235.08us  135.54us   3.09ms   97.58%
    Req/Sec     8.90k     0.91k    9.36k    98.02%
  Latency Distribution
     50%  217.00us
     75%  251.00us
     90%  283.00us
     99%    1.03ms
  178961 requests in 10.10s, 12.97MB read
Requests/sec:  17715.35
Transfer/sec:      1.28MB
```

Maybe a 5% slowdown but that's really within the statistical margin of error. I think this is good for a merge.