@nateberkopec 2019-10-14
I'm not sure the capacity optimization is worth it. Removed it and it's now benchmarking faster than master, and hello.sh benchmark on this branch is now `Requests/sec:  18001.34` vs master `Requests/sec:  16311.53`. Big response becnhmark is `Requests/sec:   2269.64` vs master's `Requests/sec:   2191.85`. So we are now even or even slightly faster than master branch 😆