@headius 2019-11-11
@nateberkopec Where are the benchmarks you ran above? I'd like to confirm there's no performance hit on JRuby and see if any tweaks might help improve things further.