@headius 2019-11-11
This is failing on JRuby because the HTTP parser ext still references the Java IOBuffer. I'll have a patch shortly.