@dentarg 2019-11-11
@headius maybe these? https://github.com/puma/puma/tree/v4.3.0/benchmarks/wrk