@headius 2019-11-11
On a simple Roda benchmark, this change appears to have no negative performance impact for JRuby. It's within the margin of error for this wrk-driven Roda benchmark. 👍